+++ controllers/doguVolumeManager_test.go
@@ -259,7 +259,7 @@ func Test_editPVCStep_Execute(t *testing.T) {
 }
 
 func Test_checkIfPVCIsResizedStep_execute(t *testing.T) {
-	t.Run("success", func(t *testing.T) {
+	t.Run("success for capacity available", func(t *testing.T) {
 		// given
 		dogu := readDoguCr(t, ldapCrBytes)
 		oldSize := dogu.Spec.Resources.DataVolumeSize
@@ -284,7 +284,41 @@ func Test_checkIfPVCIsResizedStep_execute(t *testing.T) {
 		assert.Equal(t, "Scale up", state)
 	})
 
-	t.Run("fail to get pvc ", func(t *testing.T) {
+	t.Run("success for condition FileSystemResizePending has status true", func(t *testing.T) {
+		// given
+		dogu := readDoguCr(t, ldapCrBytes)
+		oldSize := dogu.Spec.Resources.DataVolumeSize
+		defer func() {
+			dogu.Spec.Resources.DataVolumeSize = oldSize
+		}()
+		dogu.Spec.Resources.DataVolumeSize = "1Gi"
+		requests := map[corev1.ResourceName]resource.Quantity{}
+		requests[corev1.ResourceStorage] = resource.MustParse("1Gi")
+		doguPvc := &corev1.PersistentVolumeClaim{
+			ObjectMeta: *dogu.GetObjectMeta(),
+			Status: corev1.PersistentVolumeClaimStatus{
+				Conditions: []corev1.PersistentVolumeClaimCondition{
+					{Type: corev1.PersistentVolumeClaimFileSystemResizePending, Status: corev1.ConditionTrue},
+				},
+			},
+			Spec: corev1.PersistentVolumeClaimSpec{
+				Resources: corev1.ResourceRequirements{Requests: requests},
+			},
+		}
+		client := fake.NewClientBuilder().WithObjects(doguPvc).Build()
+		recorder := extMocks.NewEventRecorder(t)
+		recorder.On("Event", dogu, "Normal", "VolumeExpansion", "Wait for pvc to be resized...")
+		sut := &checkIfPVCIsResizedStep{client: client, eventRecorder: recorder}
+
+		// when
+		state, err := sut.Execute(context.TODO(), dogu)
+
+		// then
+		require.NoError(t, err)
+		assert.Equal(t, "Scale up", state)
+	})
+
+	t.Run("fail to get pvc", func(t *testing.T) {
 		// given
 		dogu := readDoguCr(t, ldapCrBytes)
 		client := fake.NewClientBuilder().Build()
@@ -321,6 +355,40 @@ func Test_checkIfPVCIsResizedStep_execute(t *testing.T) {
 		assert.Equal(t, "Wait for resize", stage)
 	})
 
+	t.Run("should return requeue error if status of condition FileSystemResizePending is false", func(t *testing.T) {
+		// given
+		dogu := readDoguCr(t, ldapCrBytes)
+		oldSize := dogu.Spec.Resources.DataVolumeSize
+		defer func() {
+			dogu.Spec.Resources.DataVolumeSize = oldSize
+		}()
+		dogu.Spec.Resources.DataVolumeSize = "1Gi"
+		requests := map[corev1.ResourceName]resource.Quantity{}
+		requests[corev1.ResourceStorage] = resource.MustParse("0.5Gi")
+		doguPvc := &corev1.PersistentVolumeClaim{
+			ObjectMeta: *dogu.GetObjectMeta(),
+			Status: corev1.PersistentVolumeClaimStatus{
+				Conditions: []corev1.PersistentVolumeClaimCondition{
+					{Type: corev1.PersistentVolumeClaimFileSystemResizePending, Status: corev1.ConditionFalse},
+				},
+			},
+			Spec: corev1.PersistentVolumeClaimSpec{
+				Resources: corev1.ResourceRequirements{Requests: requests},
+			},
+		}
+		client := fake.NewClientBuilder().WithObjects(doguPvc).Build()
+		recorder := extMocks.NewEventRecorder(t)
+		recorder.On("Event", dogu, "Normal", "VolumeExpansion", "Wait for pvc to be resized...")
+		sut := &checkIfPVCIsResizedStep{client: client, eventRecorder: recorder}
+
+		// when
+		_, err := sut.Execute(context.TODO(), dogu)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "pvc resizing is in progress")
+	})
+
 	t.Run("should return requeue error if size is not changed", func(t *testing.T) {
 		// given
 		dogu := readDoguCr(t, ldapCrBytes)
@@ -333,6 +401,40 @@ func Test_checkIfPVCIsResizedStep_execute(t *testing.T) {
 		requests[corev1.ResourceStorage] = resource.MustParse("0.5Gi")
 		doguPvc := &corev1.PersistentVolumeClaim{ObjectMeta: *dogu.GetObjectMeta(), Status: corev1.PersistentVolumeClaimStatus{Capacity: requests}, Spec: corev1.PersistentVolumeClaimSpec{
 			Resources: corev1.ResourceRequirements{Requests: requests}}}
+		client := fake.NewClientBuilder().WithObjects(doguPvc).Build()
+		recorder := extMocks.NewEventRecorder(t)
+		recorder.On("Event", dogu, "Normal", "VolumeExpansion", "Wait for pvc to be resized...")
+		sut := &checkIfPVCIsResizedStep{client: client, eventRecorder: recorder}
+
+		// when
+		_, err := sut.Execute(context.TODO(), dogu)
+
+		// then
+		require.Error(t, err)
+		assert.ErrorContains(t, err, "pvc resizing is in progress")
+	})
+
+	t.Run("should return requeue error if size is not changed and there is no condition FileSystemResizePending", func(t *testing.T) {
+		// given
+		dogu := readDoguCr(t, ldapCrBytes)
+		oldSize := dogu.Spec.Resources.DataVolumeSize
+		defer func() {
+			dogu.Spec.Resources.DataVolumeSize = oldSize
+		}()
+		dogu.Spec.Resources.DataVolumeSize = "1Gi"
+		requests := map[corev1.ResourceName]resource.Quantity{}
+		requests[corev1.ResourceStorage] = resource.MustParse("0.5Gi")
+		doguPvc := &corev1.PersistentVolumeClaim{
+			ObjectMeta: *dogu.GetObjectMeta(),
+			Status: corev1.PersistentVolumeClaimStatus{
+				Conditions: []corev1.PersistentVolumeClaimCondition{
+					{Type: corev1.PersistentVolumeClaimResizing, Status: corev1.ConditionTrue},
+				},
+			},
+			Spec: corev1.PersistentVolumeClaimSpec{
+				Resources: corev1.ResourceRequirements{Requests: requests},
+			},
+		}
 		client := fake.NewClientBuilder().WithObjects(doguPvc).Build()
 		recorder := extMocks.NewEventRecorder(t)
 		recorder.On("Event", dogu, "Normal", "VolumeExpansion", "Wait for pvc to be resized...")